+++ gettsim/synthetic_data/synthetic.py
@@ -5,13 +5,14 @@ import numpy as np
 import pandas as pd
 
 from gettsim.config import ROOT_DIR
+from gettsim.config import SUPPORTED_GROUPINGS
 from gettsim.policy_environment import _load_parameter_group_from_yaml
 
 current_year = datetime.datetime.now().year
 
 
 def append_other_hh_members(
-    df, hh_typ, n_children, age_adults, age_children, double_earner
+    df, hh_typ, n_children, age_adults, gen_female, age_children, double_earner
 ):
     """
     duplicates information from the one person already created
@@ -26,6 +27,7 @@ def append_other_hh_members(
     if hh_typ == "couple":
         adult = df.copy()
         adult["alter"] = age_adults[1]
+        adult["weiblich"] = gen_female[1]
         if not double_earner:
             adult["bruttolohn_m"] = 0
 
@@ -58,6 +60,7 @@ def create_synthetic_data(
     hh_typen=None,
     n_children=None,
     age_adults=None,
+    gen_female=None,
     age_children=None,
     baujahr=1980,
     double_earner=False,
@@ -78,6 +81,9 @@ def create_synthetic_data(
     age_adults (list of int):
         Assumed age of adult(s)
 
+    gen_female (list of bool):
+        Assumend gender of adult(s), 'False' male 'True' female
+
     age_children (list of int):
         Assumed age of children (first and second child, respectively)
 
@@ -95,7 +101,7 @@ def create_synthetic_data(
 
     kwargs:
 
-    bruttolohn_m, kapitaleink_brutto_m, eink_selbst_m, vermögen_hh (int):
+    bruttolohn_m, kapitaleink_brutto_m, eink_selbst_m, vermögen_bedürft_hh (int):
         values for income and wealth, respectively.
         only valid if heterogenous_vars is empty
     """
@@ -106,6 +112,8 @@ def create_synthetic_data(
         n_children = [0, 1, 2]
     if age_adults is None:
         age_adults = [35, 35]
+    if gen_female is None:
+        gen_female = [False, True]
     if age_children is None:
         age_children = [3, 8]
 
@@ -127,20 +135,25 @@ def create_synthetic_data(
         if (a < 0) or (type(a) != int):
             raise ValueError(f"illegal value for age: {a}")
 
+    for g in gen_female:
+        if g not in [False, True]:
+            raise ValueError("gender weiblich must be bool.")
+
     p_id_min = 0
-    hh_id_min = 0
-    tu_id_min = 0
+    group_mins = {}
+    for g in SUPPORTED_GROUPINGS:
+        group_mins[g] = 0
 
     if len(heterogeneous_vars) == 0:
         # If no heterogeneity specified,
         # just create the household types with default incomes.
         synth = create_one_set_of_households(
             p_id_min,
-            hh_id_min,
-            tu_id_min,
+            group_mins,
             hh_typen,
             n_children,
             age_adults,
+            gen_female,
             age_children,
             baujahr,
             double_earner,
@@ -159,7 +172,7 @@ def create_synthetic_data(
                 "bruttolohn_m",
                 "kapitaleink_brutto_m",
                 "eink_selbst_m",
-                "vermögen_hh",
+                "vermögen_bedürft_hh",
             ]:
                 raise ValueError(
                     f"Illegal value for variable to vary across households: {hetvar}"
@@ -170,11 +183,11 @@ def create_synthetic_data(
                         synth,
                         create_one_set_of_households(
                             p_id_min,
-                            hh_id_min,
-                            tu_id_min,
+                            group_mins,
                             hh_typen,
                             n_children,
                             age_adults,
+                            gen_female,
                             age_children,
                             baujahr,
                             double_earner,
@@ -185,8 +198,8 @@ def create_synthetic_data(
                     ]
                 )
                 p_id_min = synth["p_id"].max() + 1
-                hh_id_min = synth["hh_id"].max() + 1
-                tu_id_min = synth["tu_id"].max() + 1
+                for g in SUPPORTED_GROUPINGS:
+                    group_mins[g] = synth[f"{g}_id"].max() + 1
                 dim_counter += 1
         synth = synth.reset_index(drop=True)
     return synth
@@ -194,11 +207,11 @@ def create_synthetic_data(
 
 def create_one_set_of_households(
     p_id_min,
-    hh_id_min,
-    tu_id_min,
+    group_mins,
     hh_typen,
     n_children,
     age_adults,
+    gen_female,
     age_children,
     baujahr,
     double_earner,
@@ -214,6 +227,7 @@ def create_one_set_of_households(
         "kind",
         "bruttolohn_m",
         "alter",
+        "weiblich",
         "rentner",
         "alleinerz",
         "wohnort_ost",
@@ -246,13 +260,27 @@ def create_one_set_of_households(
         "behinderungsgrad",
         "mietstufe",
         "immobilie_baujahr",
-        "vermögen_hh",
+        "vermögen_bedürft_hh",
         "entgeltp",
         "grundr_bew_zeiten",
         "grundr_entgeltp",
         "grundr_zeiten",
         "priv_rente_m",
         "schwerbeh_g",
+        "m_pflichtbeitrag",
+        "m_freiw_beitrag",
+        "m_mutterschutz",
+        "m_arbeitsunfähig",
+        "m_krank_ab_16_bis_24",
+        "m_arbeitslos",
+        "m_ausbild_suche",
+        "m_schul_ausbild",
+        "m_geringf_beschäft",
+        "m_alg1_übergang",
+        "m_ersatzzeit",
+        "m_kind_berücks_zeit",
+        "m_pfleg_berücks_zeit",
+        "y_pflichtbeitr_ab_40",
     ]
     # Create one row per desired household
     n_rows = len(hh_typen) * len(n_children)
@@ -260,8 +288,8 @@ def create_one_set_of_households(
         columns=output_columns,
         data=np.zeros((n_rows, len(output_columns))),
     )
-    df["hh_id"] = pd.RangeIndex(n_rows) + hh_id_min
-    df["tu_id"] = pd.RangeIndex(n_rows) + tu_id_min
+    for g in group_mins:
+        df[f"{g}_id"] = pd.RangeIndex(n_rows) + group_mins[g]
 
     # Some columns require boolean type. initiate them with False
     for bool_col in [
@@ -293,6 +321,7 @@ def create_one_set_of_households(
 
     # 'Custom' initializations
     df["alter"] = age_adults[0]
+    df["weiblich"] = gen_female[0]
     df["immobilie_baujahr"] = baujahr
 
     # Household Types
@@ -316,7 +345,7 @@ def create_one_set_of_households(
     df["bruttolohn_m"] = kwargs.get("bruttolohn_m", 0)
     df["kapitaleink_brutto_m"] = kwargs.get("kapitaleink_brutto_m", 0)
     df["eink_selbst_m"] = kwargs.get("eink_selbst_m", 0)
-    df["vermögen_hh"] = kwargs.get("vermögen_hh", 0)
+    df["vermögen_bedürft_hh"] = kwargs.get("vermögen_bedürft_hh", 0)
 
     # append entries for children and partner
     for hht in hh_typen:
@@ -332,6 +361,7 @@ def create_one_set_of_households(
                         hht,
                         nch,
                         age_adults,
+                        gen_female,
                         age_children,
                         double_earner,
                     ),
@@ -368,10 +398,28 @@ def create_one_set_of_households(
     df["entgeltp"] = df["grundr_zeiten"] / 12
     df["grundr_entgeltp"] = df["entgeltp"]
 
+    # Rente Wartezeiten
+    df["m_pflichtbeitrag"] = (df["alter"] - 25).clip(lower=0) * 12
+    df["y_pflichtbeitr_ab_40"] = (df["alter"] - 40).clip(lower=0) * 12
+    df["m_freiw_beitrag"] = 5
+    df["m_ersatzzeit"] = 0
+    df["m_schul_ausbild"] = 10
+    df["m_arbeitsunfähig"] = 0
+    df["m_krank_ab_16_bis_24"] = 0
+    df["m_mutterschutz"] = 0
+    df["m_arbeitslos"] = 0
+    df["m_ausbild_suche"] = 0
+    df["m_alg1_übergang"] = 0
+    df["m_geringf_beschäft"] = 0
+    df["m_kind_berücks_zeit"] = 24
+    df["m_pfleg_berücks_zeit"] = 1
+
+    group_ids = [f"{g}_id" for g in SUPPORTED_GROUPINGS]
     df = df.reset_index()
-    df = df.sort_values(by=["hh_id", "tu_id", "index"])
+    df = df.sort_values(by=group_ids + ["index"])
     df = df.drop(columns=["index"]).reset_index(drop=True)
     df["p_id"] = df.index + p_id_min
-    df = df.sort_values(by=["hh_id", "tu_id", "p_id"])
 
-    return df[["p_id", "hh_id", "tu_id", "hh_typ"] + output_columns]
+    df = df.sort_values(by=group_ids + ["p_id"])
+
+    return df[["p_id"] + group_ids + ["hh_typ"] + output_columns]