+++ gettsim/tests/test_docs.py
@@ -9,7 +9,7 @@ from gettsim.config import TYPES_INPUT_VARIABLES
 from gettsim.functions_loader import _convert_paths_to_import_strings
 from gettsim.functions_loader import _load_functions
 from gettsim.functions_loader import load_aggregation_dict
-from gettsim.interface import rchop
+from gettsim.interface import remove_group_suffix
 from gettsim.policy_environment import load_reforms_for_date
 from gettsim.tests.utils_tests import nice_output_list_of_strings
 
@@ -80,9 +80,10 @@ def test_all_input_vars_documented(
         c
         for c in arguments
         if (c not in defined_functions)
-        and (rchop(rchop(c, "_tu"), "_hh") not in defined_functions)
+        and (remove_group_suffix(c) not in defined_functions)
         and (not c.endswith("_params"))
     ]
+
     assert not check, nice_output_list_of_strings(check)