+++ CHANGES.rst
@@ -5,6 +5,7 @@ This is a record of all past ``gettsim`` releases and what went into them in rev
 chronological order. We follow `semantic versioning <https://semver.org/>`_ and all
 releases are available on `Anaconda.org <https://anaconda.org/gettsim/gettsim>`_.
 
+
 0.5.2 —
 ------------------
 * :gh:`323` Health care contribution parameters are better aligned with law.
@@ -16,7 +17,19 @@ releases are available on `Anaconda.org <https://anaconda.org/gettsim/gettsim>`_
   date (:ghuser:`m-pannier`, :ghuser:`lillyfischer`).
 * :gh:`393` normal retirement age adjustment aligned with the rules.
   (:ghuser:`TeBackh`).
-
+* :gh:`396` Implement pension for (very) long term insured, including eligibility criteria,
+   i.e. "Wartezeiten". Implement pension for women (:ghuser:`TeBackh`).
+* :gh:`403`  Replace `Bokeh` with `plotly` for visualization.
+  (:ghuser:`effieHAN`,:ghuser:`sofyaakimova`).
+* :gh:`380` Implement an automatic type conversion of imported variables. Adding test
+  cases in `test_interface.py`. (:ghuser:`LauraGergeleit`, :ghuser:`ChristianZimpelmann`)
+* :gh:`406` Changed variable name from `vermögen_hh` to `vermögen_bedürft_hh`. Add test
+  cases for several transfers (:ghuser:`LauraGergeleit`).
+* :gh:`383` Remove ä, ö, ü from file names (:ghuser:`ChristianZimpelmann`).
+* :gh:`415` Define supported groupings in `config.py`. (:ghuser:`LauraGergeleit`,
+  :ghuser:`ChristianZimpelmann`)
+* :gh:`416` Added documentation page for Vermögensbegriff for transfers.
+  (:ghuser:`LauraGergeleit`)
 
 0.5.1 — 2022-04-21
 ------------------
@@ -64,8 +77,6 @@ releases are available on `Anaconda.org <https://anaconda.org/gettsim/gettsim>`_
   `ges_rente.yaml` (:ghuser:`hmgaudecker`, :ghuser:`ChristianZimpelmann`)
 * :gh:`356` Define functions on scalars and have them vectorised. Implement aggregation
   functions on group levels. (:ghuser:`LauraGergeleit`, :ghuser:`ChristianZimpelmann`)
-* :gh:`380` Implement an automatic type conversion of imported variables. Adding test
-  cases in `test_interface.py`. (:ghuser:`LauraGergeleit`, :ghuser:`ChristianZimpelmann`)
 
 0.4.2 — 2022-01-25
 --------------------